+++ src/components/UserInfo.js
@@ -9,7 +9,8 @@ export default class UserInfo {
     return {
       name: this.nameSelector.textContent,
       about: this.aboutSelector.textContent,
-      avatar: this.avatarSelector.src
+      avatar: this.avatarSelector.src,
+      id: ''
     }
   }
